+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 2.17.0-rc.1
+
+* Add the ability to get all annotations from an annotation manager.
+* Add `customData` field to annotations to be able to pass user data.
+
 ### 2.17.0-beta.1
 
   > [!NOTE]